- splitted regions.js in two files, one for RegionsPlugin class, one for Region class, and moved both to plugin/regions dir - added findInDirectory function in the plugin build file, so the plugins entries can be found also if in deeper directories, recursively
